Реализован ViewPager для картинки дня Apod.
